+++ packages/core/src/exports/index.ts
@@ -100,6 +100,18 @@ export {
 } from "../actions/getBackOfQueueWallet.js";
 
 export {
+    type GetSupportedTokensReturnType,
+    type GetSupportedTokensErrorType,
+    getSupportedTokens,
+} from "../actions/getSupportedTokens.js";
+
+export {
+    type GetTokenPricesReturnType,
+    type GetTokenPricesErrorType,
+    getTokenPrices,
+} from "../actions/getTokenPrices.js";
+
+export {
     type GetNetworkOrdersReturnType,
     type GetNetworkOrdersErrorType,
     getNetworkOrders,