+++ examples/in-kind-gas-sponsorship/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @renegade-fi/in-kind-gas-sponsorship-example
 
+## 0.1.27
+
+### Patch Changes
+
+- Updated dependencies
+  - @renegade-fi/core@0.9.4
+  - @renegade-fi/node@0.6.6
+  - @renegade-fi/token@0.0.22
+
 ## 0.1.26
 
 ### Patch Changes